Implement correct loading of a surveyJSON and add wait cards for ev2
The server will now read a json file with a a name that corresponds to the id contained in the route. Then its send to the client and parsed, so that the survey is loaded correctly. Also there are now different waiting cards. When completing a survey there will now be an additional card, that tells a participant what to do. Implementations are made respective to issues #41 and #42.